:truck: Rename transaction_types to types
Keep consistent naming for types
