+++ elements/lisk-transactions/test/register_delegate.spec.ts
@@ -15,7 +15,7 @@
 import { getAddressAndPublicKeyFromPassphrase } from '@liskhq/lisk-cryptography';
 import { registerDelegate } from '../src/register_delegate';
 import { DelegateAsset } from '../src/10_delegate_transaction';
-import { TransactionJSON } from '../src/transaction_types';
+import { TransactionJSON } from '../src/types';
 import * as secondSignatureReg from '../fixtures/transaction_multisignature_registration/multisignature_registration_2nd_sig_equivalent_transaction.json';
 
 describe('#registerDelegate transaction', () => {
