+++ elements/lisk-transactions/test/sign_multi_signature_transaction.spec.ts
@@ -15,7 +15,7 @@
 import { registerMultisignature } from '../src/register_multisignature_account';
 import { signMultiSignatureTransaction } from '../src/sign_multi_signature_transaction';
 import * as multisigFixture from '../fixtures/transaction_multisignature_registration/multisignature_registration_transaction.json';
-import { TransactionJSON } from '../src/transaction_types';
+import { TransactionJSON } from '../src/types';
 import { TransferTransaction } from '../src';
 
 import cloneDeep = require('lodash.clonedeep');
